requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Flask-CORS==3.0.10
python-dotenv==0.19.0
supabase>=2.0.0
numpy==1.22.4
pandas==1.5.0
scikit-learn==1.7.0
scipy==1.8.0
google-generativeai==0.3.0
Pillow==9.0.1
matplotlib==3.5.1
opencv-python==4.5.5.64
tensorflow==2.8.0
keras==2.8.0
h5py==3.6.0
pyyaml==6.0
requests==2.27.1
speechrecognition==3.8.1
protobuf==3.19.6
pytesseract==0.3.10
cryptography==3.4.7
